chore(panel): address several comments in PR
