+++ packages/react/src/components/Panel/PanelHeader.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import React, { ReactNode, forwardRef } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
@@ -7,24 +7,26 @@ export interface PanelHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
   className?: string;
 }
 
-const PanelHeader = ({
-  children,
-  className,
-  ...otherProps
-}: PanelHeaderProps) => {
-  return (
-    <div className={classNames('Panel__Header', className)} {...otherProps}>
-      {children}
-    </div>
-  );
-};
+const PanelHeader = forwardRef<HTMLDivElement, PanelHeaderProps>(
+  ({ children, className, ...otherProps }: PanelHeaderProps, ref) => {
+    return (
+      <div
+        className={classNames('Panel__Header', className)}
+        ref={ref}
+        {...otherProps}
+      >
+        {children}
+      </div>
+    );
+  }
+);
 
 PanelHeader.displayName = 'PanelHeader';
 
 PanelHeader.propTypes = {
+  //@ts-expect-error
   children: PropTypes.node.isRequired,
   className: PropTypes.string
 };
 
 export default PanelHeader;
-export { PanelHeader };
