+++ packages/styles/panel.css
@@ -14,11 +14,10 @@
 
 /* Resets aggressive Chrome user agent styles */
 .Panel > :is(.Panel__Header, .Panel__Content, .Panel__Heading) > *,
-.Panel.Panel--padding:first-child
+:is(.Panel.Panel--padding:first-child, .Panel:not(.Panel--padding):first-child)
   > :not(.Panel__Header, .Panel__Heading, .Panel__Content) {
   margin-block-start: 0;
   margin-block-end: 0;
-  color: var(--panel-content-color);
 }
 
 .Panel {
@@ -48,6 +47,7 @@
 .Panel__Content {
   padding: var(--panel-content-padding);
   margin: var(--panel-content-margin);
+  color: var(--panel-content-color);
 }
 
 .Panel__Content:not(.Panel__Content--padding) {
@@ -79,9 +79,13 @@
 
 .Panel > .Panel__Heading {
   border-bottom: var(--panel-divider-border);
+}
+
+.Panel.Panel--padding > .Panel__Heading {
   margin-bottom: var(--space-small);
 }
 
+.Panel > :is(h1, h2, h3, h4, h5, h6),
 .Panel > :is(.Panel__Header, .Panel__Heading) > :is(h1, h2, h3, h4, h5, h6),
 .Panel legend {
   font-size: var(--text-size-normal);
@@ -98,6 +102,10 @@ fieldset.Panel {
 .Panel legend {
   position: absolute;
   top: var(--space-large);
+}
+
+.Panel > .Button--primary {
+  color: white;
 }
 
 :global(.IconButton--secondary) {
